# Optimize with less memory allocation and more efficient algorithm