--- a/ext/ae_fast_decimal_formatter/extconf.rb
+++ b/ext/ae_fast_decimal_formatter/extconf.rb
@@ -1,6 +1,3 @@
 require "mkmf"
 
-abort "missing snprintf()" unless have_func "snprintf"
-abort "missing strncmp()"  unless have_func "strncmp"
-
 create_makefile "ae_fast_decimal_formatter/ae_fast_decimal_formatter"