--- a/gemfiles/ruby_2.7.5_actionview60.gemfile
+++ b/gemfiles/ruby_2.7.5_actionview60.gemfile
@@ -11,6 +11,7 @@ source "https://rubygems.org" do
   gem "rake-compiler", ">= 1", "< 2"
   gem "rspec", ">= 3.11", "< 4"
   gem "simplecov", ">= 0.21", "< 1", group: :test, require: false
+  gem "benchmark-ips"
 end
 
 gemspec path: "../"