--- a/gemfiles/ruby_3.1.0_actionview61.gemfile
+++ b/gemfiles/ruby_3.1.0_actionview61.gemfile
@@ -11,6 +11,7 @@ source "https://rubygems.org" do
   gem "rake-compiler", ">= 1", "< 2"
   gem "rspec", ">= 3.11", "< 4"
   gem "simplecov", ">= 0.21", "< 1", group: :test, require: false
+  gem "benchmark-ips"
 end
 
 gemspec path: "../"