--- a/spec/benchmark_spec.rb
+++ b/spec/benchmark_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
+require 'benchmark/ips'
 
 describe 'Benchmark performance' do
-  ITERATIONS = 100_000
   MINIMUM_EXPECTED_SPEEDUP_FACTORS__ROUNDING = 1.2
   MINIMUM_EXPECTED_SPEEDUP_FACTORS__DELIMITING = 20
 
@@ -14,37 +14,80 @@ describe 'Benchmark performance' do
     expect(current_result * minimum_speedup_factor).to be < reference_result, LOW_SPEEDUP_FACTOR_MESSAGE
   end
 
-  describe 'Rounding number' do
-    it 'comparing to Ruby string interpolation' do
-      results = Benchmark.bm do |x|
-        x.report 'Native C implementation' do
-          ITERATIONS.times { AeFastDecimalFormatter.new(355.785, 2).format }
+  describe 'formatting Fixnum' do
+    it 'should be faster in C' do
+      formatter = PureRubyDecimalFormatter.new
+      results = Benchmark.ips(quiet: true) do |x|
+        x.report 'Ruby fixnum' do
+          formatter.format(1234567890, 0)
         end
 
-        x.report 'Ruby' do
-          ITERATIONS.times { PureRubyDecimalFormatter.new(355.785, 2).format }
+        x.report 'C fixnum' do
+          AeFastDecimalFormatter.format(1234567890, 0)
         end
+
+        x.compare!
       end
 
-      native_result, ruby_result = results
-      expect_result_to_have_minimum_speedup_factor(native_result.real, ruby_result.real, MINIMUM_EXPECTED_SPEEDUP_FACTORS__ROUNDING)
+      ruby_result, c_result = results.data
+      expect_result_to_have_minimum_speedup_factor(
+        ruby_result[:cycles],
+        c_result[:cycles],
+        MINIMUM_EXPECTED_SPEEDUP_FACTORS__DELIMITING
+      )
     end
   end
 
-  describe 'Delimit large number' do
-    it 'comparing to Ruby string interpolation' do
-      results = Benchmark.bm do |x|
-        x.report 'Native C implementation' do
-          ITERATIONS.times { AeFastDecimalFormatter.new(1234567890, 0).format }
+  describe 'formatting Float' do
+    it 'should be faster in C' do
+      formatter = PureRubyDecimalFormatter.new
+      results = Benchmark.ips(quiet: true) do |x|
+        x.report 'Ruby float' do
+          formatter.format(355.785, 2)
         end
 
-        x.report 'Ruby' do
-          ITERATIONS.times { PureRubyDecimalFormatter.new(1234567890, 0).format }
+        x.report 'C float' do
+          AeFastDecimalFormatter.format(355.785, 2)
         end
+
+        x.compare!
       end
 
-      native_result, ruby_result = results
-      expect_result_to_have_minimum_speedup_factor(native_result.real, ruby_result.real, MINIMUM_EXPECTED_SPEEDUP_FACTORS__DELIMITING)
+      ruby_result, c_result = results.data
+      expect_result_to_have_minimum_speedup_factor(
+        ruby_result[:cycles],
+        c_result[:cycles],
+        MINIMUM_EXPECTED_SPEEDUP_FACTORS__ROUNDING
+      )
+    end
+  end
+
+  describe 'formatting BigDecimal' do
+    it 'should be faster in C' do
+      formatter = PureRubyDecimalFormatter.new
+      amount = BigDecimal('355.785')
+      results = Benchmark.ips(quiet: true) do |x|
+        x.report 'Ruby decimal.to_f' do
+          formatter.format(amount.to_f, 2)
+        end
+
+        x.report 'C decimal.to_f' do
+          AeFastDecimalFormatter.format(amount.to_f, 2)
+        end
+
+        x.report 'C decimal' do
+          AeFastDecimalFormatter.format(amount, 2)
+        end
+
+        x.compare!
+      end
+
+      ruby_result, c_result = results.data
+      expect_result_to_have_minimum_speedup_factor(
+        ruby_result[:cycles],
+        c_result[:cycles],
+        MINIMUM_EXPECTED_SPEEDUP_FACTORS__ROUNDING
+      )
     end
   end
 end