--- a/spec/pure_ruby_decimal_formatter.rb
+++ b/spec/pure_ruby_decimal_formatter.rb
@@ -6,12 +6,7 @@ require 'action_view/helpers/number_helper'
 class PureRubyDecimalFormatter
   include ActionView::Helpers::NumberHelper
 
-  def initialize(num, precision)
-    @num = num
-    @precision = precision
-  end
-
-  def format
-    number_with_delimiter(@num.round(@precision), delimiter: ',')
+  def format(num, precision)
+    number_with_delimiter(num.round(precision), delimiter: ',')
   end
 end