We currently format a double into a string to a certain precision.
This is done with two calls to snprintf one to generate the format
string from precision and the other to generate the number from the
double. We then copy the number string into another buffer adding
commas as needed.

Rounding is a problem and it fundamentally results from floating
point number not being able to represent precise values. For example,

148.855 -> 148.854999999999989768184605054557323455810546875
148.855.round(2) * 100 -> 14886
(148.855 * 100).round -> 14885

1234567890.12 -> 1234567890.11999988555908203125
1234567890.12.round(2) * 100 -> 123456789011.99998
(1234567890.12 * 100).round -> 123456789012

So it is very tricky to produce the right results when rounding
floating point numbers.

The current implemenation requires the use to create an instance
of AeFastDecimalFormatter and then invoke format. For example,

AeFastDecimalFormatter.new(12.213, 2).format

This is unnecessary as object allocation is slow and it produces
more garbage for the GC to worry about.

The new implementation operated on longs and is done in a single
loop that translates the indidivual digits to characters and
inserts commas as needed.

We have a class-level Ruby method wrap the c function in order to
round and massage the inputs into a long. We can thus efficiently
support BigDecimal, FixNum, and Float numbers.

Here are the benchmarks for formatting Fixnums, Floats, and BigDecimals.
The new implemenation is C2. As can be seen the new implemenation is
at least 25% faster when formatting BigDecimals and as much as 250%
faster when formatting Fixnums than the current C implemenation.

An important note is that the current implementation required BigDecimal
to be converted to a Float. This is not only slower but as mentioned
can cause rouding issues. The new implementation correctly supports
BigDecimals while still being faster.

Calculating -------------------------------------
            Ruby int     46.673k (± 6.2%) i/s -    235.850k in   5.076667s
               C int      1.552M (± 6.2%) i/s -      7.881M in   5.098310s
              C2 int      3.888M (± 6.8%) i/s -     19.709M in   5.092877s

Comparison:
              C2 int:  3888469.2 i/s
               C int:  1552058.7 i/s - 2.51x  (± 0.00) slower
            Ruby int:    46672.6 i/s - 83.31x  (± 0.00) slower

Calculating -------------------------------------
          Ruby float     61.596k (± 4.2%) i/s -    309.111k in   5.028353s
             C float      1.616M (± 5.5%) i/s -      8.062M in   5.005955s
            C2 float      3.220M (± 3.7%) i/s -     16.375M in   5.093767s

Comparison:
            C2 float:  3219522.5 i/s
             C float:  1615865.1 i/s - 1.99x  (± 0.00) slower
          Ruby float:    61595.9 i/s - 52.27x  (± 0.00) slower

Calculating -------------------------------------
   Ruby decimal.to_f     51.176k (± 9.5%) i/s -    257.184k in   5.077471s
      C decimal.to_f    571.524k (± 4.5%) i/s -      2.872M in   5.036509s
     C2 decimal.to_f    725.520k (± 3.9%) i/s -      3.676M in   5.075521s
          C2 decimal    686.669k (± 3.7%) i/s -      3.462M in   5.050310s

Comparison:
     C2 decimal.to_f:   725519.9 i/s
          C2 decimal:   686668.7 i/s - same-ish: difference falls within error
      C decimal.to_f:   571524.3 i/s - 1.27x  (± 0.00) slower
   Ruby decimal.to_f:    51176.5 i/s - 14.18x  (± 0.00) slower